requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
reportlab==4.0.7
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.2
email-validator==2.1.0.post1 